Improved the Pitch method snap

+++ src/operand_generic.py
@@ -952,16 +952,16 @@ class Pitch(Generic):
 
     def snap(self, up: bool = False) -> Self:
         scale_list: list[int] = self._get_staff() % list()
-        self_pitch: float = self.get_pitch_note()
-        pitch_offset: float = 0.0
+        self_pitch: int = self.get_pitch_note()
+        pitch_offset: int = 0
         if up:
-            pitch_step: float = 1.0
+            pitch_step: int = 1
         else:
-            pitch_step: float = -1.0
-        while scale_list[int(self_pitch + pitch_offset)] == 0:
-            pitch_offset += pitch_step
-        if pitch_offset > 0.0:
-            self += pitch_offset
+            pitch_step: int = -1
+        while scale_list[self_pitch + pitch_offset] == 0:
+            pitch_offset += float(pitch_step)
+        if pitch_offset > 0:
+            self += float(pitch_offset)
         return self
